P12 六边形/三角形密铺地图实现与测试
- 新增 world/tiling 六边形/三角形几何、环绕邻接、边界穿越
- Map/MapGenerator/Renderer/空间哈希/移动等系统适配异种密铺
- 新增 lwmap 自描述地图格式与生成/加载
- 城市多格基建地块按密铺等级形状表
- 新增 tests/test_tiling.cpp 与基线/快照更新
- 开发思路文档追加异种地图一期/二期规划

+++ src/core/MathUtil.cpp
@@ -43,7 +43,7 @@ double toScreenXf(double x, double blockSize, int panelWidth) {
     return x * blockSize + panelWidth;
 }
 
-double toScreenYf(double y, double blockSize, int mapHeight) {
+double toScreenYf(double y, double blockSize, double mapHeight) {
     return (mapHeight - y) * blockSize;
 }
 
@@ -51,7 +51,7 @@ int toScreenX(double x, double blockSize, int panelWidth) {
     return static_cast<int>(toScreenXf(x, blockSize, panelWidth));
 }
 
-int toScreenY(double y, double blockSize, int mapHeight) {
+int toScreenY(double y, double blockSize, double mapHeight) {
     return static_cast<int>(toScreenYf(y, blockSize, mapHeight));
 }
